基金购买金额改为浮点型

+++ src/api/v1/fund/schemas.py
@@ -12,7 +12,7 @@ from typing import List
 class CreateFundParams(BaseModel):
     customer_id: int = Field(..., title="客户ID")
     name: str = Field(..., title="产品名称")
-    amount: int = Field(..., title="购买金额")
+    amount: float = Field(..., title="购买金额")
     yield_rate: float = Field(..., title="收益率")
     buy_date: str = Field(..., regex=DATE_REGEX, title="购买时间")
     day_number: int = Field(..., title="产品天数")
@@ -21,7 +21,7 @@ class CreateFundParams(BaseModel):
 
 class UpdateFundParams(BaseModel):
     name: str = Field(None, title="产品名称")
-    amount: int = Field(None, title="购买金额")
+    amount: float = Field(None, title="购买金额")
     yield_rate: float = Field(None, title="收益率")
     buy_date: str = Field(None, regex=DATE_REGEX, title="购买时间")
     day_number: int = Field(None, title="产品天数")
@@ -36,7 +36,7 @@ class FetchFundsResp_Fund(BaseModel):
     id: int = Field(..., title="基金id", description="基金id")
     customer_id: int = Field(..., title="客户ID")
     name: str = Field(..., title="产品名称")
-    amount: int = Field(..., title="购买金额")
+    amount: float = Field(..., title="购买金额")
     yield_rate: float = Field(..., title="收益率")
     buy_date: str = Field(..., regex=DATE_REGEX, title="购买时间")
     day_number: int = Field(..., title="产品天数")
